Update api models
